BGInput.py / BGInput.blend - Fixed up code to simplify input handling for state checking internally; also fixed bug regarding button state checking by fixing the code up
JoyProfile.py - Fixed XBOX 360 trigger axis number

SFL.py - Changed the function names so the default bloom filters are the For-loop filters (the old "defaults" had the sample number hard-coded into the filters).

BGEAPIPlus.py - Added file to SVN. Useful to expand the existing BGE API. Currently it only adds a scenes dictionary for you to be able to get the scenes by name (i.e. logic.scenes['GUI'] as opposed to [s for s in logic.getSceneList() if s.name == 'GUI'][0]). To be documented.

TPWindow.py - Added file to SVN. Useful for making a three-piece window using BGUI to display it. A three-piece window is a window in which the corners are fit to scale, and the sides are stretched to make a window of the size you desire. To be documented.