BGInput.py - Updated to be able to properly handle joystick hat input correctly (the four directions on the hats are bit values; you're not meant to use the diagonals directly).
BGInput.blend - Updated example to use hat (d-pad) to move cube.

JoyProfile.py - I've refactored how joystick profiles work. Now, they are objects of classes, rather than dictionaries. This was done so that it would be easier to work with when you're dealing with IDEs that can do code intelligence.