BGHelper.py - Documentation for Navigate function for the CNodePath class. Navigate function now returns a dictionary of properties. This is the final update of BGHelper.py before removal in favor of a complete refactoring.
LightHelper.py - Creation of the LightManager() class to help manage lights.

SFL.py - Added a Gameboy() function to return a gameboy-esque color switching filter (the colors aren't accurate, though). Altered the Bloom() function to have a threshold to only add glow to areas that are bright enough.

Sprites.py - Removal of the obj['sprrotate'] property. Fixed bug regarding subimage selection of the sprite (sometimes it would go too high or too low, if I recall). 

TODO: Refactor to move everything into function arguments. Refactor to merge SpriteMesh() and Sprite() into a single function. If the first argument is a string, then it's mesh replacement. If it's an integer, then it's UV sliding.