Slightly rearranged how audio.py's volume level of sound playback through play_sound is calculated (instead of being + or - volume_var, it's just a random number between 0 and volume_var).
Fixed up new blur() and radial_blur() filters in filters.py. Old ones moved to blur_old and radial_blur_old().

Renamed CInputDevice to InputDevice in input.py.

Added some basic documentation to make_unique_mesh in mesh.py.

Got rid of the helper functions (Clamp() and GetDimensions()) and used the reference to the existing ones in mesh.py. Renamed functions and variables to align with PEP8.

Added a check to sprites.py to ensure that the object the sprite map is operating on exists. Otherwise, it prints out a warning (defining a custom one and throwing that would probably be the better behavior...).